REF: share methods AM/BM (#53859)
