+++ pandas/core/internals/array_manager.py
@@ -6,7 +6,6 @@ from __future__ import annotations
 import itertools
 from typing import (
     TYPE_CHECKING,
-    Any,
     Callable,
     Hashable,
     Literal,
@@ -16,10 +15,8 @@ import numpy as np
 
 from pandas._libs import (
     NaT,
-    algos as libalgos,
     lib,
 )
-from pandas.util._validators import validate_bool_kwarg
 
 from pandas.core.dtypes.astype import (
     astype_array,
@@ -261,10 +258,11 @@ class BaseArrayManager(DataManager):
         # expected "List[Union[ndarray, ExtensionArray]]"
         return type(self)(result_arrays, new_axes)  # type: ignore[arg-type]
 
-    def apply_with_block(
-        self, f, align_keys=None, swap_axis: bool = True, **kwargs
-    ) -> Self:
+    def apply_with_block(self, f, align_keys=None, **kwargs) -> Self:
         # switch axis to follow BlockManager logic
+        swap_axis = True
+        if f == "interpolate":
+            swap_axis = False
         if swap_axis and "axis" in kwargs and self.ndim == 2:
             kwargs["axis"] = 1 if kwargs["axis"] == 0 else 0
 
@@ -319,49 +317,12 @@ class BaseArrayManager(DataManager):
 
         return type(self)(result_arrays, self._axes)
 
-    def where(self, other, cond, align: bool) -> Self:
-        if align:
-            align_keys = ["other", "cond"]
-        else:
-            align_keys = ["cond"]
-            other = extract_array(other, extract_numpy=True)
-
-        return self.apply_with_block(
-            "where",
-            align_keys=align_keys,
-            other=other,
-            cond=cond,
-        )
-
-    def round(self, decimals: int, using_cow: bool = False) -> Self:
-        return self.apply_with_block("round", decimals=decimals, using_cow=using_cow)
-
     def setitem(self, indexer, value) -> Self:
         return self.apply_with_block("setitem", indexer=indexer, value=value)
-
-    def putmask(self, mask, new, align: bool = True) -> Self:
-        if align:
-            align_keys = ["new", "mask"]
-        else:
-            align_keys = ["mask"]
-            new = extract_array(new, extract_numpy=True)
-
-        return self.apply_with_block(
-            "putmask",
-            align_keys=align_keys,
-            mask=mask,
-            new=new,
-        )
 
     def diff(self, n: int) -> Self:
         assert self.ndim == 2  # caller ensures
         return self.apply(algos.diff, n=n)
-
-    def pad_or_backfill(self, **kwargs) -> Self:
-        return self.apply_with_block("pad_or_backfill", swap_axis=False, **kwargs)
-
-    def interpolate(self, **kwargs) -> Self:
-        return self.apply_with_block("interpolate", swap_axis=False, **kwargs)
 
     def shift(self, periods: int, axis: AxisInt, fill_value) -> Self:
         if fill_value is lib.no_default:
@@ -373,15 +334,6 @@ class BaseArrayManager(DataManager):
 
         return self.apply_with_block(
             "shift", periods=periods, axis=axis, fill_value=fill_value
-        )
-
-    def fillna(self, value, limit: int | None, inplace: bool, downcast) -> Self:
-        if limit is not None:
-            # Do this validation even if we go through one of the no-op paths
-            limit = libalgos.validate_limit(None, limit=limit)
-
-        return self.apply_with_block(
-            "fillna", value=value, limit=limit, inplace=inplace, downcast=downcast
         )
 
     def astype(self, dtype, copy: bool | None = False, errors: str = "raise") -> Self:
@@ -409,36 +361,6 @@ class BaseArrayManager(DataManager):
                 return arr.copy() if copy else arr
 
         return self.apply(_convert)
-
-    def replace_regex(self, **kwargs) -> Self:
-        return self.apply_with_block("_replace_regex", **kwargs)
-
-    def replace(self, to_replace, value, inplace: bool) -> Self:
-        inplace = validate_bool_kwarg(inplace, "inplace")
-        assert np.ndim(value) == 0, value
-        # TODO "replace" is right now implemented on the blocks, we should move
-        # it to general array algos so it can be reused here
-        return self.apply_with_block(
-            "replace", value=value, to_replace=to_replace, inplace=inplace
-        )
-
-    def replace_list(
-        self,
-        src_list: list[Any],
-        dest_list: list[Any],
-        inplace: bool = False,
-        regex: bool = False,
-    ) -> Self:
-        """do a list replace"""
-        inplace = validate_bool_kwarg(inplace, "inplace")
-
-        return self.apply_with_block(
-            "replace_list",
-            src_list=src_list,
-            dest_list=dest_list,
-            inplace=inplace,
-            regex=regex,
-        )
 
     def to_native_types(self, **kwargs) -> Self:
         return self.apply(to_native_types, **kwargs)
